+++ test/LinearDynamicalSystems/CommonLDS.jl
@@ -209,10 +209,9 @@ function test_initial_state_parameter_updates_common(toy_fn, ntrials=1)
     suf = StateSpaceDynamics._initialize_td_sufficient_statistics(
         Float64, lds, tsteps_per_trial
     )
-    ux_seq = [zeros(Float64, 0, size(yt, 2)) for yt in y]
-    uy_seq = [zeros(Float64, 0, size(yt, 2)) for yt in y]
-    StateSpaceDynamics._td_init_const_blocks!(ws, lds, tsteps_per_trial, y, ux_seq, uy_seq)
-    StateSpaceDynamics._aggregate_td_suff_stats!(suf, tfs, lds, ux_seq, uy_seq, y, ws)
+    data = StateSpaceDynamics.Data(lds, y)
+    StateSpaceDynamics._td_init_const_blocks!(ws, lds, data)
+    StateSpaceDynamics._aggregate_td_suff_stats!(suf, tfs, lds, data, ws)
     if lds.obs_model isa StateSpaceDynamics.GaussianObservationModel
         StateSpaceDynamics.mstep!(lds, suf, ws)
     else
@@ -279,10 +278,9 @@ function test_state_model_parameter_updates_common(toy_fn, ntrials=1)
     suf = StateSpaceDynamics._initialize_td_sufficient_statistics(
         Float64, lds, tsteps_per_trial
     )
-    ux_seq = [zeros(Float64, 0, size(yt, 2)) for yt in y]
-    uy_seq = [zeros(Float64, 0, size(yt, 2)) for yt in y]
-    StateSpaceDynamics._td_init_const_blocks!(ws, lds, tsteps_per_trial, y, ux_seq, uy_seq)
-    StateSpaceDynamics._aggregate_td_suff_stats!(suf, tfs, lds, ux_seq, uy_seq, y, ws)
+    data = StateSpaceDynamics.Data(lds, y)
+    StateSpaceDynamics._td_init_const_blocks!(ws, lds, data)
+    StateSpaceDynamics._aggregate_td_suff_stats!(suf, tfs, lds, data, ws)
     if lds.obs_model isa StateSpaceDynamics.GaussianObservationModel
         StateSpaceDynamics.mstep!(lds, suf, ws)
     else